remove redundant datasets
