+++ darts/utils/data/inference_dataset.py
@@ -1,6 +1,6 @@
 """
-Inference Dataset
------------------
+Inference Datasets
+------------------
 """
 
 import bisect
@@ -110,7 +110,7 @@ class InferenceDataset(ABC, Dataset):
         return cov_times
 
 
-class GenericInferenceDataset(InferenceDataset):
+class SequentialInferenceDataset(InferenceDataset):
     def __init__(
         self,
         series: Union[TimeSeries, Sequence[TimeSeries]],
@@ -368,299 +368,4 @@ class GenericInferenceDataset(InferenceDataset):
             sc,
             series,
             past_end + series.freq * (1 + self.output_chunk_shift),
-        )
-
-
-class PastCovariatesInferenceDataset(GenericInferenceDataset):
-    def __init__(
-        self,
-        series: Union[TimeSeries, Sequence[TimeSeries]],
-        past_covariates: Optional[Union[TimeSeries, Sequence[TimeSeries]]] = None,
-        n: int = 1,
-        stride: int = 0,
-        bounds: Optional[np.ndarray] = None,
-        input_chunk_length: int = 12,
-        output_chunk_length: int = 1,
-        output_chunk_shift: int = 0,
-        use_static_covariates: bool = True,
-    ):
-        """
-        Contains (past_target, past_covariates, future_past_covariates, static_covariates).
-
-        "future_past_covariates" are past covariates that happen to be also known in the future - those
-        are needed for forecasting with n > output_chunk_length by any model relying on past covariates.
-
-        For this reason, when n > output_chunk_length, this dataset will also emit the "future past_covariates".
-
-        Parameters
-        ----------
-        series
-            The target series that are to be predicted into the future.
-        past_covariates
-            Optionally, some past-observed covariates that are used for predictions. This argument is required
-            if the model was trained with past-observed covariates.
-        n
-            Forecast horizon: The number of time steps to predict after the end of the target series.
-        stride
-            Optionally, the number of time steps between two consecutive predictions. Can only be used together
-            with `bounds`.
-        bounds
-            Optionally, an array of shape `(n series, 2)`, with the left and right prediction start point boundaries
-            per series. The boundaries must represent the positional index of the series (0, len(series)).
-            If provided, `stride` must be `>=1`.
-        input_chunk_length
-            The length of the target series the model takes as input.
-        output_chunk_length
-            The length of the target series the model emits in output.
-        output_chunk_shift
-            Optionally, the number of steps to shift the start of the output chunk into the future.
-        use_static_covariates
-            Whether to use/include static covariate data from input series.
-        """
-
-        super().__init__(
-            series=series,
-            past_covariates=past_covariates,
-            future_covariates=None,
-            n=n,
-            stride=stride,
-            bounds=bounds,
-            input_chunk_length=input_chunk_length,
-            output_chunk_length=output_chunk_length,
-            output_chunk_shift=output_chunk_shift,
-            use_static_covariates=use_static_covariates,
-        )
-
-
-class FutureCovariatesInferenceDataset(GenericInferenceDataset):
-    def __init__(
-        self,
-        series: Union[TimeSeries, Sequence[TimeSeries]],
-        future_covariates: Optional[Union[TimeSeries, Sequence[TimeSeries]]] = None,
-        n: int = 1,
-        stride: int = 0,
-        bounds: Optional[np.ndarray] = None,
-        input_chunk_length: int = 12,
-        output_chunk_length: Optional[int] = None,
-        output_chunk_shift: int = 0,
-        use_static_covariates: bool = True,
-    ):
-        """
-        Contains (past_target, future_covariates, static_covariates) tuples
-
-        Parameters
-        ----------
-        series
-            The target series that are to be predicted into the future.
-        future_covariates
-            Optionally, some future-known covariates that are used for predictions. This argument is required
-            if the model was trained with future-known covariates.
-        n
-            Forecast horizon: The number of time steps to predict after the end of the target series.
-        stride
-            Optionally, the number of time steps between two consecutive predictions. Can only be used together
-            with `bounds`.
-        bounds
-            Optionally, an array of shape `(n series, 2)`, with the left and right prediction start point boundaries
-            per series. The boundaries must represent the positional index of the series (0, len(series)).
-            If provided, `stride` must be `>=1`.
-        input_chunk_length
-            The length of the target series the model takes as input.
-        output_chunk_length
-            Optionally, the length of the target series the model emits in output. If `None`, will use the same value
-            as `n`.
-        output_chunk_shift
-            Optionally, the number of steps to shift the start of the output chunk into the future.
-        use_static_covariates
-            Whether to use/include static covariate data from input series.
-        """
-        super().__init__(
-            series=series,
-            past_covariates=None,
-            future_covariates=future_covariates,
-            n=n,
-            stride=stride,
-            bounds=bounds,
-            input_chunk_length=input_chunk_length,
-            output_chunk_length=output_chunk_length or n,
-            output_chunk_shift=output_chunk_shift,
-            use_static_covariates=use_static_covariates,
-        )
-
-
-class DualCovariatesInferenceDataset(GenericInferenceDataset):
-    def __init__(
-        self,
-        series: Union[TimeSeries, Sequence[TimeSeries]],
-        future_covariates: Optional[Union[TimeSeries, Sequence[TimeSeries]]] = None,
-        n: int = 1,
-        stride: int = 0,
-        bounds: Optional[np.ndarray] = None,
-        input_chunk_length: int = 12,
-        output_chunk_length: int = 1,
-        output_chunk_shift: int = 0,
-        use_static_covariates: bool = True,
-    ):
-        """
-        Contains (past_target, historic_future_covariates, future_covariates, static_covariates) tuples.
-
-        Parameters
-        ----------
-        series
-            The target series that are to be predicted into the future.
-        future_covariates
-            Optionally, some future-known covariates that are used for predictions. This argument is required
-            if the model was trained with future-known covariates.
-        n
-            Forecast horizon: The number of time steps to predict after the end of the target series.
-        stride
-            Optionally, the number of time steps between two consecutive predictions. Can only be used together
-            with `bounds`.
-        bounds
-            Optionally, an array of shape `(n series, 2)`, with the left and right prediction start point boundaries
-            per series. The boundaries must represent the positional index of the series (0, len(series)).
-            If provided, `stride` must be `>=1`.
-        input_chunk_length
-            The length of the target series the model takes as input.
-        output_chunk_length
-            The length of the target series the model emits in output.
-        output_chunk_shift
-            Optionally, the number of steps to shift the start of the output chunk into the future.
-        use_static_covariates
-            Whether to use/include static covariate data from input series.
-        """
-        super().__init__(
-            series=series,
-            past_covariates=None,
-            future_covariates=future_covariates,
-            n=n,
-            stride=stride,
-            bounds=bounds,
-            input_chunk_length=input_chunk_length,
-            output_chunk_length=output_chunk_length or n,
-            output_chunk_shift=output_chunk_shift,
-            use_static_covariates=use_static_covariates,
-        )
-
-
-class MixedCovariatesInferenceDataset(GenericInferenceDataset):
-    def __init__(
-        self,
-        series: Union[TimeSeries, Sequence[TimeSeries]],
-        past_covariates: Optional[Union[TimeSeries, Sequence[TimeSeries]]] = None,
-        future_covariates: Optional[Union[TimeSeries, Sequence[TimeSeries]]] = None,
-        n: int = 1,
-        stride: int = 0,
-        bounds: Optional[np.ndarray] = None,
-        input_chunk_length: int = 12,
-        output_chunk_length: int = 1,
-        output_chunk_shift: int = 0,
-        use_static_covariates: bool = True,
-    ):
-        """
-        Contains (past_target, past_covariates, historic_future_covariates, future_covariates, future_past_covariates,
-        static_covariates)
-        tuples. "future_past_covariates" are past covariates that happen to be also known in the future - those
-        are needed for forecasting with n > output_chunk_length by any model relying on past covariates.
-
-        Parameters
-        ----------
-        series
-            The target series that are to be predicted into the future.
-        past_covariates
-            Optionally, some past-observed covariates that are used for predictions. This argument is required
-            if the model was trained with past-observed covariates.
-        future_covariates
-            Optionally, some future-known covariates that are used for predictions. This argument is required
-            if the model was trained with future-known covariates.
-        n
-            Forecast horizon: The number of time steps to predict after the end of the target series.
-        stride
-            Optionally, the number of time steps between two consecutive predictions. Can only be used together
-            with `bounds`.
-        bounds
-            Optionally, an array of shape `(n series, 2)`, with the left and right prediction start point boundaries
-            per series. The boundaries must represent the positional index of the series (0, len(series)).
-            If provided, `stride` must be `>=1`.
-        input_chunk_length
-            The length of the target series the model takes as input.
-        output_chunk_length
-            The length of the target series the model emits in output.
-        output_chunk_shift
-            Optionally, the number of steps to shift the start of the output chunk into the future.
-        use_static_covariates
-            Whether to use/include static covariate data from input series.
-        """
-        super().__init__(
-            series=series,
-            past_covariates=past_covariates,
-            future_covariates=future_covariates,
-            n=n,
-            stride=stride,
-            bounds=bounds,
-            input_chunk_length=input_chunk_length,
-            output_chunk_length=output_chunk_length or n,
-            output_chunk_shift=output_chunk_shift,
-            use_static_covariates=use_static_covariates,
-        )
-
-
-class SplitCovariatesInferenceDataset(GenericInferenceDataset):
-    def __init__(
-        self,
-        series: Union[TimeSeries, Sequence[TimeSeries]],
-        past_covariates: Optional[Union[TimeSeries, Sequence[TimeSeries]]] = None,
-        future_covariates: Optional[Union[TimeSeries, Sequence[TimeSeries]]] = None,
-        n: int = 1,
-        stride: int = 0,
-        bounds: Optional[np.ndarray] = None,
-        input_chunk_length: int = 12,
-        output_chunk_length: int = 1,
-        output_chunk_shift: int = 0,
-        use_static_covariates: bool = True,
-    ):
-        """
-        Contains (past_target, past_covariates, future_covariates, future_past_covariates, static_covariates) tuples.
-        "future_past_covariates" are past covariates that happen to be also known in the future - those
-        are needed for forecasting with n > output_chunk_length by any model relying on past covariates.
-
-        Parameters
-        ----------
-        series
-            The target series that are to be predicted into the future.
-        past_covariates
-            Optionally, some past-observed covariates that are used for predictions. This argument is required
-            if the model was trained with past-observed covariates.
-        future_covariates
-            Optionally, some future-known covariates that are used for predictions. This argument is required
-            if the model was trained with future-known covariates.
-        n
-            Forecast horizon: The number of time steps to predict after the end of the target series.
-        stride
-            Optionally, the number of time steps between two consecutive predictions. Can only be used together
-            with `bounds`.
-        bounds
-            Optionally, an array of shape `(n series, 2)`, with the left and right prediction start point boundaries
-            per series. The boundaries must represent the positional index of the series (0, len(series)).
-            If provided, `stride` must be `>=1`.
-        input_chunk_length
-            The length of the target series the model takes as input.
-        output_chunk_length
-            The length of the target series the model emits in output.
-        output_chunk_shift
-            Optionally, the number of steps to shift the start of the output chunk into the future.
-        use_static_covariates
-            Whether to use/include static covariate data from input series.
-        """
-        super().__init__(
-            series=series,
-            past_covariates=past_covariates,
-            future_covariates=future_covariates,
-            n=n,
-            stride=stride,
-            bounds=bounds,
-            input_chunk_length=input_chunk_length,
-            output_chunk_length=output_chunk_length or n,
-            output_chunk_shift=output_chunk_shift,
-            use_static_covariates=use_static_covariates,
         )
